% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.adlasso.R
\name{multi.adlasso}
\alias{multi.adlasso}
\title{Multi-split Adaptive Lasso}
\usage{
multi.adlasso(X, Y, covar.num = NULL, fold.num)
}
\arguments{
\item{X}{An input matrix, of dimension nobs x nvars.}

\item{Y}{A numeric response vector, containing nobs variables.}

\item{covar.num}{Number of covariates in model, default is NULL.}

\item{fold.num}{The number of cross validation folds.}
}
\value{
A list of two numeric objects of index of (1) selected and (2) unselected variables.
}
\description{
Multi-splitted variable selection using Adaptive Lasso
}
