% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.scad.R
\name{multi.scad}
\alias{multi.scad}
\title{Multi-split SCAD}
\usage{
multi.scad(X, Y, p.fac = NULL, fold.num)
}
\arguments{
\item{X}{An input matrix, of dimension nobs x nvars.}

\item{Y}{A numeric response vector, containing nobs variables.}

\item{p.fac}{A sequence of penalty factor applied on each variable.}

\item{fold.num}{The number of cross validation folds.}
}
\value{
A list of two numeric objects of index of (1) selected and (2) unselected variables.
}
\description{
Multi-splitted variable selection using SCAD
}
